refactor: completely overhaul components so that every manifest or helm release is a component (#293)
That gives fine control on dependencies.
Additionally changes to the component id avoids the need for a dedicated migration feature
as the old concept of components, where every manifest or helm release was coupled to an overarching component
with a unique component id, is dropped. Components are now decoupled from cue packages and can be moved to any
cue package.